refactor: extract SessionLibraryItem.swift (#665 slice A)
First slice of #665 (SessionLibrary.swift decomposition). Extracts:
- protocol SessionLibraryItem
- struct SessionLibraryEntry
- struct SessionLibrarySnapshot<Item>

Byte-identical move; SHA-256 a8db5c7bdb83153e4d1ee9b00b117744dfb8dec7ba21d3cc0ef803967285aeef
verified against origin/main:SessionLibrary.swift lines 3-93.

Reduces SessionLibrary.swift from 419 to 327 lines.

Tests: 667 (unchanged).
